backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
jinja2==3.1.4
itsdangerous==2.2.0
openpyxl==3.1.5
python-multipart==0.0.20
Pillow==11.1.0
pytesseract==0.3.13
opencv-python-headless==4.10.0.84

